karyon_core::crypto::key_pair

Trait PublicKeyExt

trait PublicKeyExt {
    // Required methods
    fn as_bytes(&self) -> &[u8] ;
    fn verify(&self, msg: &[u8], signature: &[u8]) -> Result<()>;
}

An extension trait, adding essential methods to all PublicKey types.

Required Methods

fn as_bytes(&self) -> &[u8]

fn verify(&self, msg: &[u8], signature: &[u8]) -> Result<()>

Verify a signature on a message with this public key.