Module karyon_core::event

event::EventSys implementation.

Structs

Event

An event within the EventSys.

EventListener

EventListener listens for and receives events from the EventSys.

EventSys

EventSys emits events to registered listeners based on topics.

Constants

CHANNEL_BUFFER_SIZE 🔒

Traits

EventValue

EventValueAny

EventValueTopic

Type Aliases

EventListenerID

Listeners 🔒