Struct karyon_core::event::EventSys

pub struct EventSys<T> {
    listeners: Mutex<HashMap<T, HashMap<String, HashMap<EventListenerID, Sender<Event>>>>>,
}

EventSys supports event emission to registered listeners based on topics.

Example

use karyon_core::event::{EventSys, EventValueTopic, EventValue};

 async {
    let event_sys = EventSys::new();

    #[derive(Hash, PartialEq, Eq, Debug, Clone)]
    enum Topic {
        TopicA,
        TopicB,
    }

    #[derive(Clone, Debug, PartialEq)]
    struct A(usize);

   impl EventValue for A {
        fn id() -> &'static str {
            "A"
        }
    }

    let listener = event_sys.register::<A>(&Topic::TopicA).await;

    event_sys.emit_by_topic(&Topic::TopicA, &A(3)) .await;
    let msg: A = listener.recv().await.unwrap();

    #[derive(Clone, Debug, PartialEq)]
    struct B(usize);

    impl EventValue for B {
        fn id() -> &'static str {
            "B"
        }
    }

    impl EventValueTopic for B {
        type Topic = Topic;
        fn topic() -> Self::Topic{
            Topic::TopicB
        }
    }

    let listener = event_sys.register::<B>(&Topic::TopicB).await;

    event_sys.emit(&B(3)) .await;
    let msg: B = listener.recv().await.unwrap();

    // ....
 };

Fields
listeners: Mutex<HashMap<T, HashMap<String, HashMap<EventListenerID, Sender<Event>>>>>
Implementations

impl<T> EventSys<T>where T: Hash + Eq + Debug + Clone,

pub fn new() -> ArcEventSys<T>

Creates a new EventSys

pub async fn emit<E: EventValueTopic<Topic = T> + Clone>(&self, value: &E)

Emits an event to the listeners.

The event must implement the EventValueTopic trait to indicate the topic of the event. Otherwise, you can use emit_by_topic().

pub async fn emit_by_topic<E: EventValueAny + EventValue + Clone>( &self, topic: &T, value: &E )

Emits an event to the listeners.

pub async fn register<E: EventValueAny + EventValue + Clone>( self: &Arc<Self>, topic: &T ) -> EventListener<T, E>

Registers a new event listener for the given topic.

async fn remove(&self, topic: &T, event_id: &str, listener_id: &EventListenerID)

Removes the event listener attached to the given topic.